corrected a test and split integration tests into separate files
